refactor: BottomFooterLayoutコンポーネントを再読み込み #55
